chore: moved functions to helpers
